Avoid rewriting "reg" addresses
uvmm rewrites "reg" properties in the device tree in a few places,
usually to ensure that device tree nodes have the correct size depending
on the actually mapped dataspace or similar.

This is done by reading the original reg property and then rewriting it,
only changing the size, e.g.:

   uint64_t fb_addr, fb_size;
   node.get_reg_val(0, &fb_addr, &fb_size);
   fb_size = gfb->buffer()->size();
   node.set_reg_val(fb_addr, fb_size);

This works in most cases, but results in incorrect behavior when the
device tree is using address translation using "ranges", e.g.:

    l4vmm@e000000 {
        compatible = "simple-bus";
        #address-cells = <1>;
        #size-cells = <1>;
        ranges = <0 0 0xe000000 0x01000000>;

        framebuffer@800000 {
            compatible = "simple-framebuffer";
            reg = <0x800000 0x400000>;
            l4vmm,fbcap = "fbcon";
        };
    };

In this case node.get_reg_val() returns the translated address
(0xe800000) but node.set_reg_val() will write it to the property
without translating it back. The reg property of the framebuffer is
changed to reg = <0xe800000 [real-size]>; even though only the size
was meant to be changed.

If we only want to change the size it is more reliable and more
efficient to do that in-place, rather than rewriting the entire
property. That way the original address is just kept as-is and there
is no need to resize the DT and to flush phandle/parent caches because
no node offsets change.

Fix this for the GIC and framebuffer code. There are two more usages of
node.set_reg_val() that are potentially affected:

  - vm_ram.cc uses it for /memory nodes, but those should be always
    top-level and therefore should not use address translation.
  - mmio_proxy.cc uses it in register_physmap_region(). The reg address
    is actually changed there (it's not enough to preserve the original
    value) so a different approach would be needed there.

Change-Id: I064c5de28584058cc1ecac9b230c9fc037b4bd90