amd64: run PIC as secondary IC behind the IOAPIC
A physical device sends interrupts to both, the IOAPIC and the PIC. The
OS has to take care to not unmask one line at both controllers.
We want to mimic this behavior to give the guest the choice of selecting
the preferred IC.

Since we cannot configure DT nodes to have two interrupt parents, we
use the IOAPIC as primary IC in the DT and remove the PIC from the DT.

The IOAPIC now creates the PIC device and forward the 16 legacy
interrupts to the PIC, thus mimicking the hardware behavior.

fixes ticket #CD-926.

Change-Id: Iaa06c408903478d92871fe0ee1bc123336299094