amd64: Start IOAPIC with all vectors masked
The IOAPIC specification mentions bit 48 to be set, while all other
bits in a redirection table entry are unspecified. It seems odd, because
bit 48 is in a reserved area. It would make more sense to initially set
the mask bit (16).
I suspect the upper half / lower half of the entry got confused and the
specification actually means bit 16 in the lower half register.

KVM sets just the mask bit on IOAPIC reset.

Change-Id: Id167ef5adb551988c46ac8244ddcf60408041c98