device_tree: Fix update_reg_val() error handling
I mistakenly assumed that set_prop_partial() throws an exception like
the other setprop() functions but this is not the case. Instead it
returns the error value directly. This means that update_reg_val()
does not actually catch errors at the moment.

Fix this by checking for the return value in set_prop_partial()
and use ERR() instead of throwing an exception directly.

Change-Id: I1096db87b06f980ac80c1c9384cfb79c76d27711