Fix physical address space size detection
When detecting the physical address space size preferably via the
0x80000008 extended leaf, we must actually check for the leaf presence
by querying the 0x80000000 extended function information (not by
querying the 0x00 basic information).

Only when the extended leaf is not available, we check for the 0x01
basic leaf presence and detect the physical address space size based on
it.

Also make sure the code is forward compatible by not using the strict
equality comparison against the highest extended leaf supported.

Change-Id: I8cc64048d61e34797cb8dd6b045c6f25fc838874