amd64: fix MSI destination ID computation
The MSI/-X address field encodes the destination ID bits in two parts:
bits 0-7 and bits 8-31. The bitmap used to store the address returns the
value of the upper bits of the destination ID as value starting at bit
zero. This value must be aligned to start at bit eight, following the
lower bits.

Change-Id: I47f2e50e0c0fd9b756d0d31bcb62ce7572bee1d6