amd64: refactor event recording and injection
This introduces a per core event recorder, which records events to
inject into the guest on VMentry.
Each place which generates an event, e.g. a GPF exception, creates a new
specialized Event type and adds it to the recorder.
Each event has an assigned priority, events happening during instruction
emulation, e.g. INT1, INT3, INTn, GPF, etc., have a higher priority than
events that are deliverd on instruction boundary, e.g. NMIs, IRQs.

The recorder orders all events by priority and selects the highest
priority event.
The event itself is then resposible to determine if the VMstate allows
for its injection and if so to provide the data to program the event
injection registers.

One implementation detail is noteworthy:
  * IRQs and NMIs are managed by the local APIC. Only one placeholder
    event each is recorded in the event recorder. The events hold a
    reference to the local APIC to determine which IRQ to inject.

The following interfaces become obsolete and are removed
  * inject_event wrappers for IRQ, NMI, HW events
  * reinject_event_after_vmexit

Change-Id: I5917866541164a919daf910fb7ba94cddb82e5bb