amd64: emulate debug registers
Set the kernel enforced MOV DR exit bit in the primary
processor-based execution controls, ignore writes to debug registers and
read them as zero.

For SVM set the corresponding intercepts and emulate the corresponding
accesses the same way. An access to DR8-15 raise an exception.

The hardware debug registers must not be accessible to the guest. Since
we perform no further action on debug registers writes, we ignore the
value written by the guest.

Change-Id: I9ad42107bd758629601a89e28bcf148c2159c675