aarch64 - mask FEAT_SVE & FEAT_PAuth support in feature registers
To be able to mask feature bits on arm64 this change enables the TID3
bit in the HCR, so that accesses to various feature registers trap into
uvmm. For most of these uvmm just passes the value unchanged, but in a
few registers this masks the bits relevant for SVE and PAuth support.

Change-Id: If4e459259e66397838cba810f9ac921d4f1f42b9