amd64/vmx: remove unnecessary VM exit after guest's first STI
With interrupt window exit control set the guest experiences a VM exit
when it first opens the interrupts after boot. This exit is only
necessary when there is actually an IRQ pending.
Analysis shows this is not always the case, thus leave it to the IRQ
management to enable/disable the bit on demand.

Disabling this flag at the initial boot configuration does not lead to
a IRQ injection stall, even though the IRQ mangement runs only after
the initial VMentry.

Case A) IRQ arrives after initial VMentry:
A VMexit is forced on the guest and the IRQ is recorded. The IRQ
management will enable the interrupt window.
Case B) IRQ pending before initial VMentry:
In this case the interrrupt window exit bit should be set, but is not
anymore. Nevertheless, the guest must perform multiple privileged
instructions during boot (e.g. mode change) which will lead to the one
necessary VMexit which triggers the IRQ management.

Change-Id: Iafef376059c5bda581f9220744d70b9830050b23