amd64/pit: refactor to support latch and read-back commands
The latch and read-back commands need support in each channel/counter to
transparently read the latch instead of the current count, if there is a
latched value.  The existing implementation cannot support this
transparent read operation.

The refactoring introduces the latches for counter and status to each
channel, as well as support for latch and read-back commands.
The refactoring moves most of the logic into the Channel abstraction
which leads to a simplified Pit_timer implementation of io_in/io_out.

Also, fix a bug where also channel 2 emits IRQ0 when reaching a count of
zero.

Change-Id: I38ebc03205fd96e44c43e6cdc99f4c7028c698cf