arm: always run threads in EL0 by default
On a !cpu_virt kernel, threads naturally run in EL0. We now keep the
same default for cpu_virt kernels too. Because a VMM must have access
to certain registers in EL1, add the required loader flags to be
started in EL1.

For arm64 the syscall must be overridden to use the correct
instruction. The KIP is not usable because it will use the wrong
instruction.

Change-Id: I2b949d873d541dbe16ebb124e53a258903030dcc