amd64/svm: intercept DR6 and DR7 accesses as well
For SVM the kernel enforces intercepts for all debug registers except
for DR6 and DR7, which SVM virtualizes. Uvmm intercepts these two as
well, to keep the behavior consistent with VMX.

Furthermore, this avoids any inconsistent state where DR6 and DR7 are
virtualized and guest specific, but DR0-3 are not.

Change-Id: Ia33d5e0b73e7d94fc954b2660e75b04407931bc8