pci: refactor PCI host bridge class and bus range parsing
Move the bus management to a dedicated class to make the interface
for bus management more visible. This also allows other bridges to
reuse the bus class for management of their downstream bus.

The bus range a host bridge serves is specified in the device tree
node. The first parameter of the bus-range property states the
bus number downstream of the PCI host bridge. The code did not reflect
this properly, which was not a problem since we only used bus 0.

With the refactoring of the PCI bus in a separate class, we need to
explicitly specify the PCI bus number and thus have to forward the
number given in the device tree node.

Change-Id: Id6dcd55cc3e42ec4b4b9857ee8e3a9277496ccc1