amd64: do not inject an NMI after guest executed STI
The interruptibility state reports a blocked by STI flag, if the guest
executed the STI instruction and thus interrupts are still blocked
for the following instructions.
The SDM states that this may also block NMI. I have not found anything
specifying the case(s) where NMI blocking is in effect, thus we don't
inject NMIs when this flag is set.

Change-Id: I0e5edfccba23681a8a9339c3885df40731c38c91