amd64: clear STI shadow flag on instruction emulation
Execution of STI with RFLAGS.IF = 0 blocks maskable interrupts on
the instruction boundary following its execution. In this state NMIs
may also be inhibited.  The inhibition ends after delivery of another
event (e.g., exception) or the execution of the next instruction.

This is reflected by the 'blocked by STI' flag in the guest's
interruptibility state. When we emulate an instruction, we have to
clear this flag to account for the fact that the inhibition window
has ended and to subsequently allow ourselves to inject interrupts or
NMIs into the guest.

Change-Id: Ic9e5d6cecdfda49b418e8e5054611c40e63ced79