PCI: support 32bit wide access to MSI capability at offset 0
The guest can access the first 32bit of a pass-through PCI device's
MSI capability with a 32bit wide read or write. Up to this commit, we
forwarded a read to the hardware and ignored a write.

Now, we do a combined read and write: The first 16bit access is
forwarded to the hardware and the second 16bits are read/written from
the emulation. Both results are combined to the requested 32-bit wide
access.

This commit does not provide support for arbitrary access width/offset
combinations for all fields of the MSI capability. A message warns about
this fact, if the guest performs an unsupported access.

Change-Id: I1d83a613a25b387152a95ffb2b4404bda6a235c8