amd64: support operating modes for counter 0 and 2 of PIT.
Modes 0 to 3 are now supported. Modes 4 and 5 are difficult to support
on counter two, as the emulation is not precise enough to make the
single-PIT-tick change in the output bit(5) at port 0x61 visible.

On counter 0 this would not be an issue, because all modes trigger an
interrupt to be observed by the guest.

Furthermore, counter 0 is always active as the gate is not connected.
This eliminates the hardware triggered modes 1 and 5 because gate
cannot go from low to high.
Counter 2 can observe a gate change, as its gate is connected to bit 0
of port 0x61.

Change-Id: I92336798a6eb859487e114542dee752b63acfb21