amd64/vmx: refactor and fix software interrupts/exceptions
A event record for a software exception or interrupt needs to store
the event type, the event number and instruction length. The type is
pre-defined and can therefore be a compile time constant.

Furthermore, on VMX software events need to set the VM-entry instruction
length as injection of these events advance the IP by the size of the
emulated instruction.
Since, this must be done when the event is selected for the injection,
the event record must store this information. As the event record has
only access to the Vm_state pointer a generic interface is necessary.

Change-Id: Iea3c1634868747bdfb8e76666dd444f78ca314eb