amd64: print info on vcpu resume failure
VMX and SVM implement different error signaling schemes for their
VMRESUME/VMRUN instructions. SVM signals an VMEXIT_INVALID exit via
the normal VMexit registers. VMX writes to the VM instruction error
field in the VMCS.

Therefore, we cannot handle both in the same way. This commit adds
the VM instruction error field to the error output in case uvmm
observes a failure in vcpu_resume_commit().

For SVM there is a stub for the future to print information for this
error case.

Change-Id: I3806f5545a2b916289e253fa3731bc839bfbd082