amd64/cpuid: precisely report the extended state size
The kernel limits the XCR0 features available to the guest to the
x87, SSE, AVX and AVX512 states. uvmm cannot detect this limitation,
thus we have to fall back on the hard-coded approach to provide a
consistent view on XCR0 through the XGETBV and CPUID instruction to
the guest.

As a result we also have to compute the XSAVE state area sizes in the
CPUID emulation. As the CPUID instruction is rather expensive and we
have to evaluate multiple sub-leafs, we build a cache and answer
subsequent calls from said cache.

see ticket #CD-989

This is based on a contribution from Nicolas Frinker of genua, who
reported the issue with the CPUID emulation for the xsave feature set.

Change-Id: Ieabeba50d834847e10978097d12b706cc7d720a1