amd64/ioapic: add support for legacy IRQs to IOAPIC
For legacy IRQs we need to be able to send EOI signals to the device
sending a level-triggered IRQ. Therefore, the IOAPIC needs to register
an EOI handler per IRQ line at the local APICs indicated by the
redirection table programming.

When sending a level-triggered IRQ onwards to the local APIC, the
remote IRR bit must indicate that the IOAPIC is waiting for an EOI
for this IRQ vector.
Furthermore, we must buffer the state of a level-triggered IRQ, while
the IRQ line is masked.

Change-Id: Ie9b5d2d5c8312e50c64e9b5ef67e82518f5f6803