amd64/lapic: allow to rebind the same EOI handler
Linux writes the RTE of the IOAPIC twice with the same value
at least during boot. There is no need to fail at the rebind
of the EOI handler at the LAPIC as long as the handler address
and thus the handler stays the same.

Also improve the error output incase the EOI handler is overwritten with
a different handler.
The error output now indicates which handler is already bound and
which new handler shall be set before failing.

Change-Id: Idce6a517e8e9b754a746848649ccc03f498c6509