virtio_device_proxy fixes
Add support for IOMMUs and make sure that the cache is flushed when
changing the queue notify number in the config space (Linux maps this
uncached).

Also make sure the receive window has enough space for an incoming cap
from the clients. It may be already queued in the kernel and waiting for
it to be delivered.

Change-Id: Ief20aaec33f534abb3f88edcd5c73ce8e36bd2cd