x86: Fix usage of setup_real_mode for resume
This was broken in I8040e08a146f87d351906c2ef8477ce51d43b8ad, it seems
unnecessary to be that strict about the entry ip. For resume, we also
want to start in real mode, but at the wakeup vector that the guest
provided on suspend. The wakeup vector does not conform to the
0x000VV000 alignment requirements for the SIPI startup vector.

After reset, the CPU operates on the CS base address, which is a
hidden part of the CS register. This base address can be used until the
CS register is loaded the next time. Then the base address is computed
normally by shifting the value of the CS register four bits to the left.
We use the CS base address and keep CS and RIP at zero to have the guest
core start at the requested instruction.

AMD and Intel CPUs behave the same in this regard.

Co-authored-by: Philipp Eppelt <philipp.eppelt@kernkonzept.com>
Change-Id: I01e3f9b8db10786bbe23d3315dfdc276d85f760a