Annotate MMIO regions
When two MMIO region overlap we print an error message telling us
various data points. However, they are usually hard to interpret,
because it is not clear which devices are generating those conflict.

For some devices we use RTTI to generate a readable name. But for others
we simply don't, because they all use the same class under the hood
(namely Ds_handler).

To make this error message more useful, switch to an explicit
scheme where every MMIO device has to provide its name by itself. This
has the advantage that we now can print proper device names even for
devices which use the Ds_handler implementation, but also that we get
proper human readable strings (and not C++ mangled ones).

This makes it a hard requirement for every subclass of Mmio_device to
implement dev_name().

Change-Id: Ibd7a593af09e40353b13dd541d4c3a6ced70eb39