fix offset for loading a file to guest RAM
The Ram_ds can load a file to the guest RAM. It either uses the
dataspace's copy_in() interface or falls back to do the memcpy itself.

The computation of the destination address for this memcpy is wrong:
local_start references the start of the managed slice of the backing
dataspace, meaning, the offset within the backing dataspace is already
included.
Therefore, the destination of the memcpy operation must soly take the
offset within guest RAM into account.

Change-Id: Ia9d994a31b57482ed30ebcfad0a59780a22b98ec