amd64: ignore INIT-deassert IPIs
INIT-deassert IPIs are only valid in P6 and Pentium CPUs. The last
Pentium architecture CPU was Pentium D in 2005. We thus ignore
INIT-deassert IPIs, if the guest OS sends these, as a modern CPU would
do.

See Intel SDM:
  11.4.7.4 Local APIC State After It Receives an INIT-Deassert IPI

This is based on a contribution from Nicolas Frinker of genua.

Change-Id: I8f88b479550ed6450ac3cdd11858d490e727bbdc