amd64: unify PAT & MTRR handling for SVM & VMX
Both architectures behave the same with regards to PAT and MTRR MSRs.
The PAT MSR is fully virtualized in the respective extended vCPU
states.

The MTRR MSRs use the same bit-structure and the same set of registers
exist on SVM and VMX. Thus, we implement a generic handler which handles
read and write request such that linux is accepting the MTRR support and
also enables PAT usage.

This also fixes the uvmm behavior wrt malformed PAT MSR values.

Change-Id: Ib0d089a17f39e3d4233f19c008e01efe9eb44ac4