amd64/vmx: handle only guest EFER and PAT
The VM-exit controls provide a switch to load the host PAT and EFER.
The kernel enforces these bits to be set. This brings uvmm in sync with
the kernel to not suggest behavior we cannot influence.

Furthermore, we do not need to save PAT and EFER on VMexit, because
we emulate writes to both registers. Thus, the VMX fields are always
up to date.

Change-Id: I4a9e2e4132f760dcb74dc2315067495d68099c36