arm: add kernel interrupt forwarding support
Utilize the direct IRQ forwarding support in the kernel if possible.
This does only affect hardware SPI interrupts so far. As this is an
optional feature, gracefully fallback to the regular injection in uvmm.

A special caveat are migrations of interrupts that are active. This is
not supported by the kernel but uvmm must instead detach, wait for the
guest to issue the EOI and re-establish the vCPU binding.

Requires-Change-Id: I53e103dd23dc9db0711fbc60908e56726fb4ec13
Requires-Change-Id: I91aca9e1d70104ca18760207810febe1822ca08e
Change-Id: I191be9bcc3f5b9fe521405201fc6a6e9e2239a76