Add virtio device config change support
Forward any access to the virtio device config space to the task
actually handling the virtio device code. It uses the same cmd mechanism
also used for the status/queue configure code. The guest expect that all
the required actions when writing into this space are executed after the
mmio write returns. Therefore this needs to block.

Change-Id: Idf772f3d78572e28f563f8fd8e1aeacfde016e73
Requires-Change-Id: I7455b0f31826b0b182da23206f9d59b40d7b85a2