arm: use direct vtimer vIRQ forwarding
Fiasco optionally allows to forward the generic timer virtual timer
interrupt directly to the guest. Configure the vCPU state accordingly.
Uvmm still needs to cope with the possibility that the feature is not
supported by Fiasco. In this case the existing code paths are triggered
and the vIRQ is injected through the uvmm GIC emulation. Otherwise, if
Fiasco handles the vtimer vIRQ, uvmm will just never receive a
guest_ppi(1) vCPU upcall.

Requires-Change-Id: I79b12e7a8f1ba9ebe10e0e34f12403022f74f301
Requires-Change-Id: Id23c61e6ef6db4266e9e40879267161394fdc212
Change-Id: I1cc397fc0dfc4b482fcec18ded964b84c58e6640