Fix psci_system_resume()
I6f1695a2c02b3fcf673ff302bcca886401b13f29 changed the way the
instruction pointer is handled in an exception. Uvmm now sees the
instruction pointer of the trapping HVC instruction and has to
change the value to skip over the instruction.

But at least one vm_call() method also changed the instruction to
set it to a value provided by the user - Psci_device::vm_call().
Adding 4 onto this value to skip the trapping HVC instruction leads
to errors in the guest.

Therefore we move the handling of the instruction pointer to the
beginning of handle_smccc_call()

Change-Id: I58732240cb3b36ada10828387a2fef7612998f37