virtio_proxy: support no-MMU systems
The address of the virtio config page MMIO mapping depends on the
actual location and cannot be chosen. Warn that performance will suffer
if the device tree specifies an address that is different from the
actual one.

As an extension, we allow a virtio,mmio device tree node without a
"reg" property. The "reg" property will be added with the actual
address at runtime. This way, Linux can run without any performance
impact because the config page can be mapped read-only to the guest.

Change-Id: Ib2666fe1e53ef2101c61678943554b1ac544142b