generic IRQ-based mechanism for CPU specific events
The mechanism is split into a generic part and an event specific part.
The generic part handles all interaction of an IRQ event with the vCPU
registry and the registration of the IRQ object.
The event specific part contains the action that shall be executed.

The first user is the stop event to stop execution of the given CPU
device. To allow for handling in the generic CPU device, this introduces
a new pure virtual function 'stop()'.
Each architecture specific CPU device must implement this function.
Furthermore, it is architecture specific, when the IRQ is armed and
disarmed. The architecture specific CPU devices must implement this.

Change-Id: I943368217c31f9c5ccfe119ac31608970fe9ad7c