Ro_ds_mapper_t: fall back to emulation on no-MMU systems
On no-MMU systems, the dataspace can only be mapped at the physical
location in the guest. If this is not possible we can fall back to
emulation for read instructions, albeit with reduced performance,
though.

On MMU systems any mapping should be possible. If it fails, we still
want to bail out because something is fundamentally broken then.

Change-Id: I15f5a280e566ba41c32dc6fd1732f6d2b0ef18f1