arm: stop other cores via IRQ instead of ex_regs exception
The ex_regs exception can lead to the invocation of the uvmm exception
handler and thus an unhandled exception error terminating uvmm.
This happens, when the ex_regs exception hits while the thread is either
executing uvmm code (vCPU kernel mode) or executing kernel code and
returns to uvmm.

To avoid all this, this commit switches to IRQ-based stop events based
on the recently introduced generic mechanism.

Change-Id: I5381b1cb6d9056a1cbf25da84366e6b290615905