amd64: refactor CPU state and activity state handling
This change refactors the whole CPU state handling and brings decision
for state change into one spot in the VM-handling loop.
This introduces two new states: Stopped, Halted.

The CPU is only in sleeping state after it was created. After being
rescheduled it leaves this state and never returns to it.
The Stopped stated represents now the part of the Sleeping state that
doesn't need rescheduling of the vCPU thread, e.g. on Shutdown or
Suspend.

The CPU is in Halted state when the guest executes the 'hlt'
instruction. It remains in this state, when no event is pending or the
injection of all pending events is blocked.

To allow for CPU state change in the VMexit handling, the responsible
function interfaces now takes the pointer to the CPU device.

Furthermore, we need to clear all state when switching from Init to
Running state, including undelivered, but pending events. Otherwise,
these are picked up in event injection.

Change-Id: If88bd44fa78e044d57c2d656139f7df27c3805ec