amd64: change CPU dev state only from affected vCPU thread
The virtual local APIC did changes to a remote CPU dev state on INIT
and SIPI signals. This makes it difficult to reason about the current
CPU dev state on the core itself, because there is no forced
synchronization point and it also allows for programming error.

This changes switches the change-remote-state+IRQ sequence to the use
of a specific IRQ for a specific state change. The IRQ handler - which
is executed on the target core - then changes the CPU dev state
locally. Thus, the CPU dev state is only written by the vCPU thread
itself.
This clear interface with a clear point in time when a CPU dev state
change can occur allows us to build a less complex evaluation
algorithm to determine the currently expected behavior of the vCPU.

Change-Id: Ibab66fc0dd14defd14ff9ef5fc4cb7372f5c4977