Add notify queue support in cmd mode
Make the queue notification synchronous. If we are running in cmd
mode, we also need to execute the queue notification in a synchronous
fashion, like we do with the status/queue config calls.

The reason is that a client expect the queue notification itself (not
the handling of that queue) is being completely executed when the mmio
write to that register returns to the guest. In some cases a guest may
write queue notification quickly right after each other and we would
lose the queue index information if we handle it asynchronous.

Change-Id: Ifb5cb4dac572e92ec47a452544739eaf017d072c
Requires-Change-Id: I04dd2f2b4dc5c072b766e4788daf38597460a3a5