Enable suspend/resume with firmware
Uvmm started out with the explicit goal of running guest Linux
in a VM on L4Re with minimal emulation. Therefore it directly
implements the Linux boot protocol and loads Linux without
requiring any form of firmware.

In the event of a system suspend, Uvmm takes the role of the
firmware and interprets the FACS->FirmwareWakingVector to
directly resume Linux at the therein specified address.

With the addition of a firmware such as EFI, this approach is
no longer valid because it is up to the firmware to resume the
guest OS.

This change implements support for S3 suspend/resume with a
firmware. In the event of a S3 suspend/resume cycle, Uvmm
restarts the bootprocessor at the reset vector. From there
the firmware boots. It looks into the CMOS to find out if
it is a resume operation. If that is case, the firmware will
restore the system state and subsequently start the guest
at the FACS->FirmwareWakingVector.

In detail, this change:
- adds a qemu-fw-if table to signal S3 support to the firmware
- sets the resume address to the reset vector when Uvmm is
  configured to boot a firmware
- adds support for CMOS memory at the RTC.
- sets the required bits in the CMOS before resume

Change-Id: Ie11b5653f46f270e1527a284f187659dae1887d4