amd64: rtc write support
This change implements write support for the emulated real
time clock (RTC) on amd64. You may now use 'hwclock -w' in
Linux to set it.

If the capability connecting Uvmm to the rtc server has write
permission, Uvmm will also set the time there.

If the rtc server has an event where its wallclock time
changes, it sends out a notification to all its clients so
that they can retrieve the new time.

This change also adds support for this update notification.

If there is no connection to an rtc server, this change at
least allows the guest to program the virtual RTC. This RTC
will tick just like a normal RTC, but it will not be able to
track wallclock time for example when the machine is suspended
and resumed.

Change-Id: Ibb1c052360c3d66edb8da6272f083e25cd265fc1