Add simple GICD_CTRL disable implementation
GICD_CTLR enables the forwarding of pending interrupts from the
Distributor to the CPU interfaces. Here we provide a simple
implementation of the EnableGrp* bits by checking whether any of
the group bits is set.

To provide a full implementation we we would need to check the
EnableGrp bit associated with a pending irq.

Change-Id: I43f0bd2b0367efd586c7b2cf839c069351b26a7b