amd64: remove cpu state check for INIT on startup IPI
Checking a remote cores state after sending a state change IRQ gives
no guarantees that the remote state change already happened. Thus,
reading the remote state can contain the old or the new state.

This removes the flawed state check in the INIT-SIPI-SIPI IPI sequence.
If the state change IRQs are received (or issued) and processed in
wrong order, the target CPU's state machine must deal with this.

Change-Id: I2b3a49f7721f97efedd2bb652731cee514fe4259