[API] make 'num_ds' factory argument optional
The 'num_ds' factory argument is mandatory and doesn't use the common
"key=value" scheme. That makes it also position dependent and not
compatible with e.g. the virtio net P2P link.

This change makes the argument optional by providing a default of '2'
and also changes it to use our "key=value" scheme. The key is 'ds-max'
which makes it compatible with the virtio net P2P link.

Change-Id: I2ee97f5c86e7a7c77725ad87ec6f9987e97e7543