amd64: handle real mode exceptions
In case the vCPU still operates in real mode, the exception delivery
differs. Exceptions do not push an error code on the stack. VM-entry
fails if we try to do so.
For hardware exceptions, check if the current operating mode is
real mode and use a new real-mode exception type.

Change-Id: If3d5cf9738ec069794af0612d04a631c2a4fb6d2