Implement OpenBSD Boot Protocol
Add an additional OpenBSD kernel loader and automatically detect an
OpenBSD kernel binary by its OPENBSD_RANDOM ELF header section.

In case an OpenBSD kernel is loaded, use the OpenBSD boot protocol to
directly boot this kernel.
This mimics the OpenBSD bootloader.

Additionally, we need the size of the kernel image, so add
infrastructure for getting this value.

Furthermore, the initial GUEST_RSP register is setup dynamically now,
because the boot parameters are communicated via this register.

Change-Id: I72dfce8ffbe9d4f25d972d3927723c9c717992bc