cfi: proactively map flash memory to avoid instruction decoding
The Cfi flash memory must only be mapped RO to the guest, if the flash
memory command is read_array. Otherwise, read accesses must be emulated
depending on the current flash command.
When the guest writes the read_array command, we do now map the memory
instead of waiting for the next faulting read.

This saves one VMexit and avoids the need to decode instructions
doing a memory read, like `cmpb   $0x2,0x37(%rdi)`.

Change-Id: I2139cceaa9320603eac5bf535ac78a669cc05be4