Compile qemu-fw-if by default
With this change we select CONFIG_UVMM_QEMU_FW_IF be default. The
reasoning is that we want to ensure firmware boot works by default.

Previously the user could forget to compile the feature which would
result in the VM failing to boot, which is very hard to debug if the
user does not have deep knowledge about the system (and an EFI image
with debug symbols).

Change-Id: I1d93144bf5e0789ca6e0e43beb0f18c922e0c97a