amd64: handle triple fault exit of vCPUs
On a triple fault exit(VMX) or shutdown exit(SVM), the core enters
shutdown mode, meaning it waits for an INIT signal.
In principle, the guest can recover the affected core, unless the
last/only running core experiences a triple fault.
In this case uvmm shuts down.

This adds a check for the number of running cores to the CPU dev's
state handling.

fixes ticket #CD-1277

Change-Id: I511a8912fbf05949bd6be56ed39ffd6076da1a16