amd64: act only on first SIPI in INIT-SIPI-SIPI sequence
AMD CPUs don't need the second SIPI to boot up correctly, Intel requests
this to happen in the manual.

We need to filter one SIPI, because the SIPI handling code within uvmm
does a cross-core write in the vCPU state. Which leads to undefined
behavior, when the vCPU already executes.
We cannot easily get rid of this behavior, so keep the number of SIPIs
after INIT at one.

The code now only acts on the INIT-SIPI part of the initialization
sequence. All SIPIs after the first are ignored. To achieve this, only
an INIT signal zeros the internal SIPI counter.

This change also eliminates another possibility for misbehavior: The old
code acted on every second SIPI without the need for an INIT IPI to
have happen. So an INIT-SIPI-SIPI-SIPI-SIPI would have initialized the
target vCPU twice.

Change-Id: I2c7539db5b683e70991554bfdd76409e9b3e1598