virtio-pci: improve MSIx config/queue vector handling
The Virtio PCI connector stored the MSI-X vector numbers for config and
queue notifications. The config vector was never communicated to the
actual virtio device, the queue vector was only passed to the virtio
device, when the virtio device had MSI-X enabled at the time of enabling
the queue.

The buffering of the values and the restriction on when to write the
vector number to the device is unnecessary. We can directly write these
numbers into the virtio queue config and rely on the device not to use
them before the guest enables the device.
The guest is responsible to complete device setup wrt the MSI-X vectors
before enabling the device.

Furthermore, the queue config vector is now passed to the device as
well.

This is based on a contribution by Nicolas Frinker of genua.

Change-Id: I3e8639059c9324c45695615eb82cdf10302beed5