amd64: preallocate memory for event records
To avoid dynamic memory de-/allocation during VMexit handling this
commit introduces a preallocation and management mechansim. The
Event_record instances are now allocated from this memory pool and
the memory is returned on destruction.
In case our pool runs out of memory, we fall back to dynamic memory
allocation. A warning is printed once, when dynamic memory allocation
is necessary, to allow the user to tune the preallocation size.

The ownership of Event_records is still with the Event_manager, which
handles the object destruction.

fixes ticket #CD-1233.

Requires-Change-Id: I34d7ad05e039c88be14cff2fb04a6629d8f64231
Change-Id: Ie26adc12bfa5684f4344a12067586e639767d75c