fix double allocation of PCI device IDs per virtIO proxy device
Each virtIO proxy device allocated a device ID on the PCI bus, which it
used to determine the MSI-X route and then it registered itself on the
PCI bus, which lead to a second allocation of a device ID. The second
device ID is the ID registered on the PCI bus and the address of the
device from the guest point of view.

Provide the already allocated device ID at registration on the PCI bus
and avoid the second device ID allocation.

Change-Id: I48a4eec6f1b51c4675865fd00b20ea264f856fc6