amd64: Fix in_real_mode()
The original implementation of in_real_mode() was wrong. When the CR0.PE
bit is set, the CPU is running in protected mode (i.e. not running in
real mode). However, the function returned true in that case.

While at it, move the method to the implementation file for consistency
with other methods that will come in the future and that check various
CPU modes.

Change-Id: If8000c6e8650a73638cd7e5977978fa276c62508