vmm.lua: create no scheduler proxy without prio and cpus given
The last iteration of set_sched() introduced the new behavior that a
scheduler proxy is always created, even when it is not wanted. The user
of vm_start() could not influence this anymore.

With this change, set_sched() does not create a scheduler proxy if
both input parameters are nil. This gives the control back to the user
of vm_start().
Furthermore, this also documents all cases and behaviors.

Change-Id: I73f2bab9349a577a533874d8e00b82162d5fd1b9