Add --dt-mod commandline parameter
This switch allows for small device tree modification at runtime. It
will add/modify the given property before the device tree is parsed by
uvmm. Strings ("str"), Bools ("bool") and Ints ("u32", "u64") are
supported. The format is dt-path/dt-property=type:val. For example to
disable a node, you could use something like this:

 -m/bus1/device1/status=str:disabled

The default type is Strings("str") and if the "=..." part is missing it
defaults to bool.

Change-Id: I6f83d48f51c90d8331f8b72c1f1ee8be552f9be4