Introduce Timer::ready() virtual method
The Timer::ready() virtual method can be overridden by a timer device to
start the timer operation (i.e. enqueue the first timeout). This is
usually not possible to do in the constructor of the device object
because the timer (clocks) infrastructure is not ready yet at that
moment.

Note: The Timer::ready() method is currently only ever executed on
AMD64. The actual implementation of timers on other platforms is future
work.

Change-Id: I9e11eb72fe57a9b846351953e6d558008ab33fb1