Add support for framebuffers with active refresh
If the framebuffer does not have the auto-refresh feature, it needs to
be actively refreshed by the client.

This is currently done by a fixed periodic timer. The timer is running
at 30 Hz by default which is the bare minimum for a pleasant visual
experience. The refresh rate can be overridden by the l4vmm,refresh_rate
device tree property.

Since there is currently no detailed information about the activity of
the VM with respect to the framebuffer available, the output is
currently always refreshed in its entirety. There are optimization
possibilities available in the future.

Change-Id: I6a6adc21e0c4ec63aad90b3bd67d6b2b2c973d28