Introduce fw_cfg item support
This adds the l4vmm,items property of the qemu_fw_if device as the
configuration data and can be used to pass arbitrary data from the host
to the VM.

The supported format (compatible with the format of the QEMU's "-fw_cfg"
command-line argument) is "[name=]name,string=string" where "name" is
the name of the configuration entry and "string" is the contents of the
configuration entry. The "name=" label is optional, the "string=" label
is mandatory.

The other QEMU format (i.e. "[name=]name,file=file") is not supported.

Change-Id: I8026e32e70d2c2a7e616f650b963870420108f75